Fixes circular dependency check failures
Previously, if you used the rails runner to prevent reloading the rails
server and you attempted to build the same bundled asset twice,
sprockets would fail with a circular dependency check error.

It turns out that sprockets caches bundled assets in memory and doesn't
clear it's circular dependency checking array, this commit goes in and
clears that array to prevent that check from failing.